fix the crash issue when the drive becomes inaccessible.
nvme_mi_ctrl instance will be created in initialize() when we can
retrieve its information from the drive. if the drive is inaccessible,
the instance wont' be created. at later on, get_log NVMe-MI command
needs this instance as the arguement to query drive SMART and then it
will crash the application.

This commit will fix it.

Fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-8256

Signed-off-by: Mars Yang <maryang@nvidia.com>