fixed the long response from drive
There are up to 4 drivers in the  same i2c bus andi each NVMe-MI request packets will
be sent out and  wait for the response sequece sequentially. The next
reuqet command should be sent out after the response is recieved.
Otherwise, many requests will be queued up and sent out the request very late.

The solution is serializing NVMe-MI commands , sepecially more drives in the
system.

The other fix is the esteimated sanitize time is not reported by drive
and  the default timerout is 30 seconds for current drives we are using.
Tested:
issue Drive Sanitize and check the progress in Task
/redfish/v1/Systems/System_0/Storage/1/Drives/201/Actions/Drive.SecureErase -d '{"SanitizationType":"BlockErase"}'

Signed-off-by: MarsYang <maryang@nvidia.com>