handle sdbus exceptions properly
When the nvme manager receives a InterfacesRemoved signal, it always
tries to read the signal message to check whether a MCTP-EID was
removed or not even if it's not a nvme-mi device.

However, if the bounded interface was terminated too quick before
message read, an SdBusError exception would be thrown from sdbusplus,
which caused unexpected SIGABRT/core-dump since there was no proper
exception handling. This usually happens for the MCTP-EIDs that can
be discovered but no service dependency that nvidia-nvme-manager
can rely on [e.g, Requires=, BindsTo or After=], such as mctp-i2c5-*
services used for CX-7 telemetry.

Here we implemented a simple try-catch to process SdBusError since
it shouldn't be handled as SIGABRT.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5054516
Signed-off-by: Chester Lin <chesterl@nvidia.com>